requirements.txt
asgiref==3.4.1
Django==3.2.11
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
